fix(errors): unify error boundaries, prevent duplicate toasts
